feat:rendered the total movies API
