+++ src/components/MovieLists/index.tsx
@@ -1,24 +1,33 @@
+import { toast } from 'react-toastify';
+
 import { useGetMovieList } from '../../apis/Movie';
 import { MovieListCard } from '../MovieListCard';
 
+import './styles.css';
+
 export const MovieList = () => {
-  const { data: movieList, isLoading, isError } = useGetMovieList();
+  const { data: movies, isLoading, isError } = useGetMovieList();
 
   if (isLoading) {
-    return <p>Loading movies...</p>;
+    toast.success('Loading Movies..');
   }
+
   if (isError) {
-    return <p>Failed to load movies. Please try again later.</p>;
+    toast.error('Failed to load movies.Please try again');
   }
-  if (!movieList || movieList.length === 0) return <p>No movies found.</p>;
+  if (!movies || movies.length === 0) return <p>No movies found.</p>;
 
   return (
     <>
-      <h2>Movie List</h2>
-      <div className="movie-list-container">
-        {movieList.map(movie => (
-          <MovieListCard key={movie.id} {...movie} />
-        ))}
+      <h2 className="now-playing-movie-lists-title">
+        Enjoy and Explore....
+      </h2>
+      <div className="now-playing-movie-container">
+        <div className="now-playing-movie-list-container">
+          {movies.map(movie => (
+            <MovieListCard key={movie.id} {...movie} />
+          ))}
+        </div>
       </div>
     </>
   );
