+++ src/components/StarRating/styles.css
@@ -1,0 +1,3 @@
+.star-rating {
+  margin-top: -19px;
+}
