+++ src/pages/Home/styles.css
@@ -1,4 +1,3 @@
 .home-body {
-  background-color: #333230;
   color: #fff;
 }
